debug weixin xin info.

+++ app/src/main/java/com/lovejob/view/_money/fragments_sendmoney/SendOriWork.java
@@ -201,7 +201,7 @@ public class SendOriWork extends BaseFragment {
         }
         String s = tvSendWorkOriPrice.getText().toString();
         String s1 =s.substring(0,s.length()-2);
-        int money = Integer.parseInt(s1);
+        Double money = Double.parseDouble(s1);
         if (money<=0){
             Utils.showToast(context, "支付金额不能小于0元");
             return;
